fix(cli): trim overlong session names to fit unix socket path limit
Long session names plus a long os.tmpdir() prefix (e.g. /var/folders/...
on macOS) pushed the cli daemon's unix socket path past sun_path's 104-byte
limit, and listen() failed with EINVAL. makeSocketPath now sanitizes the
name for fs use and hashes the middle when the full path would overflow.

Fixes https://github.com/microsoft/playwright/issues/40878
